requirements.txt
Pillow>=8.2